Reconcile the caching layer's request counts with the branches beneath it
Restacking the caching layer onto the verification fixes below it left
fifteen failing examples in the spec files those branches own. One was a
real break; the rest were fixtures whose request counts predate caching.

The rename of the nested-call slot to #outside_called_tool_definition
reached the Streamable HTTP dispatcher but not the plain HTTP one, which
still named the method that no longer exists. Every SSE-framed message
plain HTTP had to route -- a notification, a legacy server's ping, a
server-initiated request it must refuse -- raised NoMethodError inside the
response parse, so the request that carried it failed. It calls the method
the mixin actually defines again. The mixin was also included twice, and
the dispatcher's comment survived the rename in both its old and new
wording; both are tidied.

The rest is fixture drift. On a 2026-07-28 server an absent ttlMs means
"assume 0", so a list the server does not bound is stale the moment it
arrives and every later access re-reads it: a tools/call re-derives its
Mcp-Param-* headers from a list it fetches again, and a list_changed
notification has no cache left to drop. Seven examples counted requests
against a server that sends no bound, and asserted the counts of a
transport that kept its lists for the life of the connection.

Each of them now bounds the list its subject does not concern -- the
recovery composition's tools/list, the HeaderMismatch refresh sequence,
the paginated list, the three lists a change notification invalidates,
and the list a session-less modern server serves -- so that what the
example counts is the recovery, the refresh or the invalidation it names,
and not the caching layer underneath it. The counts themselves are
unchanged, and still fail if a bounded result stops being served fresh.
