+++ lib/mcp_client/json_rpc_common.rb
@@ -13,6 +13,8 @@ require_relative 'subscription_support'
 require_relative 'input_round_trips'
 require_relative 'result_caching'
 require_relative 'request_metadata'
+require_relative 'round_trip_marker'
+require_relative 'result_completeness'
 
 module MCPClient
   # Shared retry/backoff logic for JSON-RPC transports
@@ -20,6 +22,8 @@ module MCPClient
     include Envelopes
     include ErrorBodies
     include InputWaits
+    include RoundTripMarker
+    include ResultCompleteness
     include SubscriptionSupport
     include InputRoundTrips
     include ResultCaching
@@ -828,40 +832,6 @@ module MCPClient
       {}
     end
 
-    # Project a payload out of a result that has to be finished. This client
-    # recognizes InputRequiredResult (resultType "input_required") but does
-    # not drive multi round-trip requests yet, so an operation that would
-    # extract a field from it — and so drop the server's inputRequests and
-    # opaque requestState — surfaces it instead of presenting an unfinished
-    # answer as an empty successful one. The whole result rides on the
-    # error's `data`, so a host can still drive the round trip itself.
-    #
-    # Every field projector goes through this, not just resources/read: MRTR
-    # permits an unfinished result on exactly tools/call, prompts/get and
-    # resources/read, so one on a list or a completion is malformed anyway —
-    # and reading it as a finished empty page would drop a whole page of a
-    # paginated list, or a completion, without a word. tools/call and
-    # prompts/get return the entire result and so never project.
-    # An unfinished answer is the InputRequired condition and is reported as
-    # such, the same way #reject_unfulfillable_input_required! reports one
-    # that reaches the response parser; any other discriminator this client
-    # cannot carry through is an invalid result.
-    # @param result [Object] the JSON-RPC result
-    # @param method [String] the request method, for the message
-    # @return [Object] the result, when it is complete
-    # @raise [MCPClient::Errors::InputRequiredError] when it is unfinished
-    # @raise [MCPClient::Errors::InvalidResultError] when it is neither
-    def require_complete_result!(result, method)
-      type = MCPClient::JsonRpcCommon.result_type(result)
-      return result if type == 'complete'
-
-      message = "#{method} answered with resultType #{type.to_s[0, 64].inspect}, which this " \
-                'client cannot carry through'
-      raise MCPClient::Errors::InputRequiredError.new(message, data: result) if type == 'input_required'
-
-      raise MCPClient::Errors::InvalidResultError.new("Invalid result: #{message}", data: result)
-    end
-
     # Which request field mirrors into the Mcp-Name header (MCP 2026-07-28
     # Streamable HTTP "Standard Request Headers"; the tasks extension adds
     # taskId routing for its methods).
@@ -1021,26 +991,6 @@ module MCPClient
       e.request_params ||= params
       e.transport ||= self
       raise
-    end
-
-    # Whether the request this thread last resolved went through a multi
-    # round-trip retry. The marker is thread-local: a transport serves
-    # concurrent requests, and a result that depended on input responses
-    # MUST NOT be cached even if another request completed meanwhile.
-    # @return [Boolean]
-    def last_result_from_round_trip?
-      Thread.current[round_trip_marker_key] == true
-    end
-
-    # @param flag [Boolean]
-    # @return [void]
-    def mark_round_trip_result(flag)
-      Thread.current[round_trip_marker_key] = flag
-    end
-
-    # @return [Symbol] the thread-local key of this transport's round-trip marker
-    def round_trip_marker_key
-      :"mcp_client_round_trip_#{object_id}"
     end
 
     # Notifications the 2026-07-28 revision removed; never written to a
